Lock Travis to previous image to fix libpq issue
The December 1st, 2016 update added PostgreSQL 9.5 and 9.6 to the build
image, including a 9.6 libpq by default. Previously, PostgreSQL 9.4 and
its libpq were all that was present. Though our scripts always install
PostgreSQL 9.5 or 9.6, they did not upgrade libpq.

This meant we always used the system default libpq (previously 9.4),
though we did actually get the "right" version of psql (née csql).

With 9.6, libpq's default handling of CONTEXT changed: it is hidden
unless the level is ERROR or higher. We addressed this ourselves using
the SHOW_CONTEXT variable (by setting "always" in pg_regress_multi): in
9.5, this is ignored (and unneeded), in 9.6, it ensures old behavior is
preserved.

However, this only applies in builds where the libpq version matches
the psql version. If, for instance, the psql version is 9.5 and the
libpq version is 9.6, psql won't know to do with that variable, so the
underlying 9.6 libpq behavior will remain, i.e. only showing CONTEXT in
ERROR conditions or higher.

This is now precisely the case for our Travis 9.5 builds. They use the
9.5 psql binary with the system's new libpq (9.6). For now, we're going
to just slap a "group: deprecated" line in our Travis configuration; in
time we'll probably want to just use the "latest" psql for all builds,
no matter what (our hardcoding of the path was due to the obsolete csql
requirement; now that we use stock psql, we can probably go with the
postgresql-common pg_wrapper-linked psql, which selects the highest psql
version currently installed).

See the following link (which promises the above will suffice "for at
least a month") for more information:
https://docs.travis-ci.com/user/build-environment-updates/2016-12-01/